GFS: bugfix for FSIndex object returning directory as a JSON string
We save the directory content as a JSON string inside MongoDB
because of limitations in saving a JSON object containing a dot
in the key. To workaround this issue, we have a post init hook
for parsing the JSON string into an object right after loading
from DB, and a pre validate hook for stringifying the object
into a JSON string before we save into the DB.

The bug happens when an object does not exist in the DB and we
create a default index - the `Model.create` Mongoose API does
not trigger the post init hook, resulting in the directory
content being returned as a JSON string, instead of an object.